Changed PrecursorInfo::charge to int Removed cast to int in SpectrumList_PrecursorRecalculator
Former-commit-id: 301517491ba496ddf9657039485956a1c03e3f6d

+++ pwiz/analysis/spectrum_processing/PrecursorRecalculator.hpp
@@ -40,7 +40,7 @@ class PWIZ_API_DECL PrecursorRecalculator
     {
         double mz;
         double intensity;
-        double charge;
+        int charge;
         double score;
 
         PrecursorInfo() : mz(0), intensity(0), charge(0), score(0) {}
